Run a full tclean for the _post image for each solint, and use that as a start for the next solint

+++ selfcal_helpers.py
@@ -7,7 +7,7 @@ def tclean_wrapper(vis, imagename, band_properties,band,telescope='undefined',sc
                    cycleniter = 300, uvtaper = [], savemodel = 'none',gridder='standard', sidelobethreshold=3.0,smoothfactor=1.0,noisethreshold=5.0,\
                    lownoisethreshold=1.5,parallel=False,nterms=1,cyclefactor=3,uvrange='',threshold='0.0Jy',phasecenter='',\
                    startmodel='',pblimit=0.1,pbmask=0.1,field='',datacolumn='',spw='',obstype='single-point', \
-                   savemodel_only=False):
+                   savemodel_only=False, resume=False):
     """
     Wrapper for tclean with keywords set to values desired for the Large Program imaging
     See the CASA 6.1.1 documentation for tclean to get the definitions of all the parameters
@@ -74,8 +74,9 @@ def tclean_wrapper(vis, imagename, band_properties,band,telescope='undefined',sc
     if gridder=='mosaic' and startmodel!='':
        parallel=False
     if not savemodel_only:
-        for ext in ['.image*', '.mask', '.model*', '.pb*', '.psf*', '.residual*', '.sumwt*','.gridwt*']:
-            os.system('rm -rf '+ imagename + ext)
+        if not resume:
+            for ext in ['.image*', '.mask', '.model*', '.pb*', '.psf*', '.residual*', '.sumwt*','.gridwt*']:
+                os.system('rm -rf '+ imagename + ext)
         tclean(vis= vis, 
                imagename = imagename, 
                field=field,
